virttest.qemu_vm: Avoid crash on VM cleanup
It might happen that some of the files intended for cleanup in VM are
None (due to non-finished operations). This causes all following tests
to fail as VM is stored in ENV. This patch checks if f is anything.

+++ virttest/qemu_vm.py
@@ -2629,7 +2629,8 @@ class VM(virt_vm.BaseVM):
 
         for f in file_list:
             try:
-                os.unlink(f)
+                if f:
+                    os.unlink(f)
             except OSError:
                 pass
